Added AStarPathFinder to Player
Minimal implementation to test the code.

+++ core/src/com/kiriost/game/gameobject/map/IMapObserver.java
@@ -4,5 +4,5 @@ package com.kiriost.game.gameobject.map;
  * Created by kiriost on 08/04/16.
  */
 public interface IMapObserver {
-    void terrainClicked(float x, float y);
+    void terrainClicked(int x, int y);
 }
